+++ src/css/custom.css
@@ -1190,9 +1190,3 @@ code {
 	padding-right: 0.25rem;
 	font-family: var(--ifm-font-family-mono);
 }
-
-@utility scrollbar-hidden {
-	&::-webkit-scrollbar {
-		display: none;
-	}
-}
